refactor: move components to uikit
