+++ src/__tests__/Button.test.tsx
@@ -1,7 +1,7 @@
+import { Button } from '@/shared/ui-kit/Button';
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { describe, it, expect, vi } from 'vitest';
-import { Button } from '../shared/ui/kit/Button';
 
 describe('Button', () => {
   it('is a button with a text', () => {
