+++ src/components/ThemeToggler.tsx
@@ -1,5 +1,5 @@
+import { Button } from '@/shared/ui-kit/Button';
 import { useThemeContext } from '../shared/model/context/useThemeContext';
-import { Button } from '../shared/ui/kit/Button';
 
 export function ThemeToggler() {
   const { theme, toggleTheme } = useThemeContext();
