+++ src/shared/ui-kit/Button.tsx
@@ -4,13 +4,16 @@ export const Button = ({
   content,
   className = '',
   onClick,
+  isDisabled = false,
 }: {
   content: string;
   className?: string;
+  isDisabled?: boolean;
   onClick: () => void;
 }) => {
   return (
     <button
+      disabled={isDisabled}
       className={cn(
         `cursor-pointer rounded bg-[var(--color-primary)] px-4 py-2 text-sm text-[var(--color-surface)] hover:bg-[var(--color-primary-hover)]`,
         className
